mark Cook Profile as Favorite

+++ application/modals/address.js
@@ -74,10 +74,10 @@ module.exports = function (sequelize, DataTypes) {
             allowNull: false
         },
         latitude: {
-            type: DataTypes.DECIMAL(12, 9)
+            type: DataTypes.FLOAT
         },
         longitude: {
-            type: DataTypes.DECIMAL(12, 9)
+            type: DataTypes.FLOAT
         },
         updated_at: DataTypes.DATE,
         deleted_at: DataTypes.DATE
